adding a valid_url for metrotvnews
